+++ forui/lib/src/widgets/date_field/calendar/calendar_date_field.dart
@@ -187,7 +187,7 @@ class _CalendarDatePickerState extends _FDateFieldState<_CalendarDateField> {
       builder: (state) => FTextField(
         control: .managed(controller: _textController),
         focusNode: _focus,
-        style: style.textFieldStyle,
+        style: style.fieldStyle,
         textAlign: widget.textAlign,
         textAlignVertical: widget.textAlignVertical,
         textDirection: widget.textDirection,